Refactor alarms dropdown into AlarmsList
